requirements.txt
faiss-cpu
numpy
torch
sentence-transformers
pypdf
pdfplumber
pytesseract
pdf2image
groq
openai
